+++ CHANGELOG.md
@@ -17,6 +17,9 @@ If the change isn't user-facing but still relevant enough for a changelog entry,
 * (internal)? scope: short description (#pr, @author)
 -->
 
+### Changed
+* resource/anxcloud_ip_address: reserve an available address based on filters (#163, @anx-mschaefer)
+
 ## [0.6.2] - 2024-05-28
 
 ### Added